edged.go: Don't restart when container exited
see issue #1364.
When sandbox container is started but not ready, PLEG relist will
generate a container die event. Func syncLoopIteration will add the pod to
restart queue. This is unexpected compared to kubelet's op of deleting
container of the pod.
Latest container will be kept in default in kubelet.
This is a lazy way of doing to ignore the die event.
By the way, the restart logic is probmatic because the app container is
not restarted.

Signed-off-by: Faicker Mo <faicker.mo@gmail.com>